# backend/openapi.yaml
openapi: "3.0.3"
info:
  title: 智能用药App API
  description: 智能用药App后端API接口文档 - 支持慢性病患者和老年人的用药管理
  version: "1.0.0"
  contact:
    name: 后端开发团队
    email: backend@smartmed.com
servers:
  - url: http://localhost:3000/api
    description: 本地开发环境
  - url: https://api.smartmed.com/api
    description: 生产环境
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 使用JWT Bearer Token进行认证，通过登录接口获取
  schemas:
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 5
        hasMore:
          type: boolean
          example: true
...
    Medicine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "阿莫西林胶囊"
        genericName:
          type: string
          example: "阿莫西林"
        category:
          type: string
          example: "抗生素"
        dosageForm:
          type: string
          example: "胶囊剂"
        specification:
          type: string
          example: "0.5g*24粒"
        manufacturer:
          type: string
          example: "华北制药股份有限公司"
        barcode:
          type: string
          example: "6901234567890"
        indications:
          type: string
        usageDosage:
          type: string
        contraindications:
          type: string
        sideEffects:
          type: string
        precautions:
          type: string
        drugInteractions:
          type: string
        plainLanguageExplanation:
          type: string
    InteractionCheckRequest:
      type: object
      required:
        - medicineIds
      properties:
        medicineIds:
          type: array
          minItems: 2
          maxItems: 20
          items:
            type: string
            format: uuid
    InteractionResult:
      type: object
      properties:
        interactions:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              severity:
                type: string
                enum: [high, medium, low]
              severityLevel:
                type: integer
              description:
                type: string
              mechanism:
                type: string
              clinicalManagement:
                type: string
              symptoms:
                type: string
              medicineAName:
                type: string
              medicineBName:
                type: string
        summary:
          type: object
          properties:
            total:
              type: integer
            high:
              type: integer
            medium:
              type: integer
            low:
              type: integer

    # ========== Medication Plan ==========
    ScheduleItem:
      type: object
      required:
        - time
        - daysOfWeek
      properties:
        time:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          example: "08:00"
        daysOfWeek:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
          example: [1, 2, 3, 4, 5, 6, 7]
    CreatePlanRequest:
      type: object
      required:
        - medicineName
        - dosageValue
        - dosageUnit
        - frequencyType
        - schedule
      properties:
        medicineId:
          type: string
          format: uuid
        medicineName:
          type: string
          maxLength: 200
        medicineSpecification:
          type: string
        dosageForm:
          type: string
        dosageValue:
          type: number
          example: 1
        dosageUnit:
          type: string
          enum: [片, 粒, ml, mg, g, 袋, 支, 瓶]
        dosageDescription:
          type: string
          example: "每次1片"
        frequencyType:
          type: string
          enum: [daily, every_n_hours, specific_days, as_needed]
        frequencyInterval:
          type: integer
          minimum: 1
          maximum: 24
        frequencyTimesPerDay:
          type: integer
          minimum: 1
          maximum: 10
        schedule:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleItem'
          minItems: 1
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        stockQuantity:
          type: number
        stockUnit:
          type: string
        notes:
          type: string
          maxLength: 500
...
    TodaySchedule:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: array
          items:
            type: object
        completed:
          type: array
          items:
            type: object
        missed:
          type: array
          items:
            type: object
        progress:
          type: integer
          example: 60

    # ========== Check-in ==========
    CreateCheckinRequest:
      type: object
      required:
        - planId
        - scheduledDate
        - scheduledTime
        - status
      properties:
        planId:
          type: string
          format: uuid
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
        status:
          type: string
          enum: [taken, skipped, late]
        note:
          type: string
          maxLength: 500
        isMakeup:
          type: boolean
          default: false
        makeupNote:
          type: string
...
    CreateGuardianRequest:
      type: object
      required:
        - patientId
        - relationship
      properties:
        patientId:
          type: string
          format: uuid
        relationship:
          type: string
          enum: [spouse, child, parent, sibling, relative, caregiver, other]
        notes:
          type: string
    ConfirmGuardianRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [accept, reject]
    GuardianPatient:
      type: object
      properties:
        patientId:
          type: string
          format: uuid
        patientName:
          type: string
        relationship:
          type: string
        todaySummary:
          type: object
          properties:
            total:
              type: integer
            completed:
              type: integer
            missed:
              type: integer
    GuardianReport:
      type: object
      properties:
        patient:
          type: object
        relationship:
          type: string
        summary:
          type: object
          properties:
            totalDoses:
              type: integer
            completedDoses:
              type: integer
            missedDoses:
              type: integer
            complianceRate:
              type: number
        trend:
          type: array
          items:
            type: object

    # ========== Notification ==========
...
    RegisterDeviceRequest:
      type: object
      required:
        - deviceToken
        - platform
      properties:
        deviceToken:
          type: string
        platform:
          type: string
          enum: [ios, android, web]
        deviceName:
          type: string
paths:
...
  /medicines:
    get:
      tags: [药品信息]
      summary: 搜索药品
      description: 按药品名称、通用名搜索药品，支持模糊搜索
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
          description: 搜索关键词
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Medicine'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /medicines/{id}:
    get:
      tags: [药品信息]
      summary: 获取药品详情
      description: 获取药品完整信息，包括说明书
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Medicine'
  /medicines/barcode/{code}:
    get:
      tags: [药品信息]
      summary: 条码查询药品
      description: 通过药品包装条码查询药品信息
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
          description: 药品条码
      responses:
        '200':
          description: 成功
  /medicines/interactions:
    post:
      tags: [药品信息]
      summary: 药物相互作用检查
      description: 检查多种药物同时服用的相互作用风险
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionCheckRequest'
      responses:
        '200':
          description: 检查完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/InteractionResult'

  # ========== Plans ==========
  /plans:
    get:
      tags: [用药计划]
      summary: 获取用药计划列表
      security:
        - BearerAuth: []
      parameters:
        - name: includeInactive
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 成功
    post:
      tags: [用药计划]
      summary: 创建用药计划
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '201':
          description: 用药计划已创建
  /plans/today:
    get:
      tags: [用药计划]
      summary: 获取今日用药安排
      description: 获取用户今日待服药、已完成、已漏服的列表及进度
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TodaySchedule'
  /plans/{id}:
    get:
      tags: [用药计划]
      summary: 获取用药计划详情
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
    put:
      tags: [用药计划]
      summary: 更新用药计划
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 用药计划已更新
    delete:
      tags: [用药计划]
      summary: 删除用药计划
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 用药计划已删除

  # ========== Check-ins ==========
  /checkins:
    post:
      tags: [用药打卡]
      summary: 服药打卡
      description: 记录服药打卡，支持正常服药、补服、跳过
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckinRequest'
      responses:
        '201':
          description: 打卡成功
    get:
      tags: [用药打卡]
      summary: 获取打卡记录
      description: 按日期范围获取打卡记录列表
      security:
        - BearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: 成功
  /checkins/{id}/undo:
    post:
      tags: [用药打卡]
      summary: 撤销打卡
      description: 5分钟内可撤销打卡记录
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 打卡已撤销
  /checkins/calendar:
    get:
      tags: [用药打卡]
      summary: 获取日历视图数据
      description: 按年月获取每日打卡统计
      security:
        - BearerAuth: []
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 成功

  # ========== Prescriptions ==========
...
  /guardians:
    post:
      tags: [家人监护]
      summary: 发送监护邀请
      description: 创建监护关系，向被监护人发送邀请
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGuardianRequest'
      responses:
        '201':
          description: 监护邀请已发送
  /guardians/patients:
    get:
      tags: [家人监护]
      summary: 获取被监护人列表
      description: 监护人查看所有被监护人的用药摘要
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GuardianPatient'
  /guardians/my-guardians:
    get:
      tags: [家人监护]
      summary: 获取我的监护人列表
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
  /guardians/invitations:
    get:
      tags: [家人监护]
      summary: 获取待确认的监护邀请
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
  /guardians/{patientId}/confirm:
    put:
      tags: [家人监护]
      summary: 确认/拒绝监护邀请
      security:
        - BearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmGuardianRequest'
      responses:
        '200':
          description: 操作成功
  /guardians/{patientId}:
    delete:
      tags: [家人监护]
      summary: 解除监护关系
      security:
        - BearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 监护关系已解除
  /guardians/{patientId}/report:
    get:
      tags: [家人监护]
      summary: 获取被监护人用药报告
      description: 获取指定时间范围内的用药依从率报告
      security:
        - BearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GuardianReport'

  # ========== Notifications ==========
  /notifications:
    get:
      tags: [通知管理]
      summary: 获取通知列表
      security:
        - BearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: 成功
  /notifications/unread-count:
    get:
      tags: [通知管理]
      summary: 获取未读通知数量
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
  /notifications/{id}/read:
    put:
      tags: [通知管理]
      summary: 标记通知为已读
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 已标记为已读
  /notifications/read-all:
    put:
      tags: [通知管理]
      summary: 全部标记为已读
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 全部标记为已读
  /notifications/device:
    post:
      tags: [通知管理]
      summary: 注册设备推送令牌
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '200':
          description: 设备已注册

  # ========== Health ==========
...